+++ examples/tutorials/14_change_color.py
@@ -1,0 +1,14 @@
+from easygraphics import *
+from easygraphics.dialog import *
+
+init_graph(600, 400)
+
+while is_run():
+    if has_mouse_msg():
+        x, y, type, buttons = get_mouse_msg()
+        if type == MouseMessageType.PRESS_MESSAGE:
+            color = get_color(get_background_color())
+            set_background_color(color)
+    delay_fps(60)
+
+close_graph()