+++ examples/tutorials/15_music_play.py
@@ -1,0 +1,8 @@
+from easygraphics.music import *
+from easygraphics.dialog import *
+
+# For legal reasons please prepare a music file by yourself
+load_music("test.mp3")
+play_music()
+show_message("ok!")
+close_music()